original code of the Martinez CSG algorithm
The code I based on when writing polyclip-go package.

The code in cageo141/ is, IIRC, the original code from Martinez et al.,
as published on the original microsite about the algorithm, by authors.

The code in as3polyclip/ is an ActionScript3 port of the code (that's
what I actually based on, as it's more readable).

The files in polygons/ are some datafiles which were distributed
together with the C code.

The paper.pdf file is the paper as it was published on the microsite.